💄 UI: 꿀팁 상세 컴포넌트 추가 #76
